menu_bind_state struct definitions need to be moved over to menu_input.h
